# swagger.yaml
openapi: 3.0.0
info:
  title: Legal Advice
  contact: {}
  version: '1.0'
servers:
- url: https://legal-advice-1812.onrender.com/api/v1
  variables: {}
- url: http://localhost:1812/api/v1
  variables: {}
paths:
...
  /info/criminalDictionary/{id}:
    delete:
      tags:
      - Criminal Dictionary
      summary: Remove Criminal Dictionary
      operationId: RemoveCriminalDictionary
      parameters:
      - name: id
        in: path
        description: ''
        required: true
        style: simple
        schema:
          type: string
      responses:
        '200':
          description: ''
          headers: {}
      deprecated: false
      security:
      - bearer: []
    get:
      tags:
      - Criminal Dictionary
      summary: Get Specific Criminal Dictionary
      operationId: GetSpecificCriminalDictionary
      parameters:
      - name: id
        in: path
        description: ''
        required: true
        style: simple
        schema:
          type: string
      responses:
        '200':
          description: ''
          headers: {}
      deprecated: false
  /info/criminalDictionary:
    post:
      tags:
      - Criminal Dictionary
      summary: Add Criminal Dictionary
      operationId: AddCriminalDictionary
      parameters: []
      requestBody:
        description: ''
        content:
          application/json:
            schema:
              allOf:
              - $ref: '#/components/schemas/AddCriminalDictionaryRequest'
              - example:
                  title: Test
                  decs: ' Test Test Test Test Test Test Test Test Test '
            example:
              title: Test
              decs: ' Test Test Test Test Test Test Test Test Test '
        required: true
      responses:
        '200':
          description: ''
          headers: {}
      deprecated: false
      security:
      - bearer: []
    get:
      tags:
      - Criminal Dictionary
      summary: Get All Criminal Dictionary
      operationId: GetAllCriminalDictionary
      parameters: []
      responses:
        '200':
          description: ''
          headers: {}
      deprecated: false
  /news:
    post:
      tags:
      - News
      summary: Add New
      operationId: AddNew
      parameters: []
      requestBody:
        content:
          multipart/form-data:
            encoding: {}
            schema:
              required:
              - image
              - title
              - desc
              type: object
              properties:
                image:
                  type: string
                  format: binary
                title:
                  type: string
                  example: Test
                desc:
                  type: string
                  example: 'Tesr Tesr Tesr Tesr Tesr Tesr Tesr Tesr Tesr Tesr Tesr '
        required: false
      responses:
        '200':
          description: ''
          headers: {}
      deprecated: false
      security:
      - bearer: []
    get:
      tags:
      - News
      summary: Get All News
      operationId: GetAllNews
      parameters: []
      responses:
        '200':
          description: ''
          headers: {}
      deprecated: false
  /news/{id}:
    patch:
      tags:
      - News
      summary: Update New
      operationId: UpdateNew
      parameters:
      - name: id
        in: path
        description: ''
        required: true
        style: simple
        schema:
          type: string
      requestBody:
        description: ''
        content:
          application/json:
            schema:
              allOf:
              - $ref: '#/components/schemas/UpdateNewRequest'
              - example:
                  title: 'Test '
            example:
              title: 'Test '
        required: true
      responses:
        '200':
          description: ''
          headers: {}
      deprecated: false
      security:
      - bearer: []
    get:
      tags:
      - News
      summary: Get Specific New
      operationId: GetSpecificNew
      parameters:
      - name: id
        in: path
        description: ''
        required: true
        style: simple
        schema:
          type: string
      responses:
        '200':
          description: ''
          headers: {}
      deprecated: false
    delete:
      tags:
      - News
      summary: Delete Specific New
      operationId: DeleteSpecificNew
      parameters:
      - name: id
        in: path
        description: ''
        required: true
        style: simple
        schema:
          type: string
      responses:
        '200':
          description: ''
          headers: {}
      deprecated: false
      security:
      - bearer: []
  /messages:
    get:
      tags:
      - Message
      summary: Get All Messages
      operationId: GetAllMessages
      parameters: []
      responses:
        '200':
          description: ''
          headers: {}
      deprecated: false
      security:
      - bearer: []
    post:
      tags:
      - Message
      summary: Send Message To Lawyer
      operationId: SendMessageToLawyer
      parameters: []
      requestBody:
        content:
          multipart/form-data:
            encoding: {}
            schema:
              required:
              - text
              - image
              - file
              type: object
              properties:
                text:
                  type: string
                  example: Test 2
                image:
                  type: string
                  format: binary
                file:
                  type: string
                  format: binary
        required: false
      responses:
        '200':
          description: ''
          headers: {}
      deprecated: false
      security:
      - bearer: []
  /messages/{id}:
    get:
      tags:
      - Message
      summary: Get Specific Message
      operationId: GetSpecificMessage
      parameters:
      - name: id
        in: path
        description: ''
        required: true
        style: simple
        schema:
          type: string
      responses:
        '200':
          description: ''
          headers: {}
      deprecated: false
      security:
      - bearer: []
    patch:
      tags:
      - Message
      summary: Update Specific Message
      operationId: UpdateSpecificMessage
      parameters:
      - name: id
        in: path
        description: ''
        required: true
        style: simple
        schema:
          type: string
      requestBody:
        content:
          application/x-www-form-urlencoded:
            encoding: {}
            schema:
              required:
              - text
              type: object
              properties:
                text:
                  type: string
                  example: Updated Test
        required: false
      responses:
        '200':
          description: ''
          headers: {}
      deprecated: false
      security:
      - bearer: []
    delete:
      tags:
      - Message
      summary: Delete Specific Message
      operationId: DeleteSpecificMessage
      parameters:
      - name: id
        in: path
        description: ''
        required: true
        style: simple
        schema:
          type: string
      responses:
        '200':
          description: ''
          headers: {}
      deprecated: false
      security:
      - bearer: []
  /messages/me:
    get:
      tags:
      - Message
      summary: Get My Message
      operationId: GetMyMessage
      parameters: []
      responses:
        '200':
          description: ''
          headers: {}
      deprecated: false
      security:
      - bearer: []
...
  /orders:
    post:
      tags:
      - Order
      summary: Add Order
      operationId: AddOrder
      parameters: []
      requestBody:
        description: ''
        content:
          application/json:
            schema:
              allOf:
              - $ref: '#/components/schemas/AddOrderRequest'
              - example:
                  reason: Test
                  desc: Test
                  phone: '01544604708'
                  date: 2012-7
            example:
              reason: Test
              desc: Test
              phone: '01544604708'
              date: 2012-7
        required: true
      responses:
        '200':
          description: ''
          headers: {}
      deprecated: false
      security:
      - bearer: []
    get:
      tags:
      - Order
      summary: Get All Orders
      operationId: GetAllOrders
      parameters: []
      responses:
        '200':
          description: ''
          headers: {}
      deprecated: false
      security:
      - bearer: []
  /orders/{id}:
    get:
      tags:
      - Order
      summary: Get Specific Order
      operationId: GetSpecificOrder
      parameters:
      - name: id
        in: path
        description: ''
        required: true
        style: simple
        schema:
          type: string
      responses:
        '200':
          description: ''
          headers: {}
      deprecated: false
      security:
      - bearer: []
    delete:
      tags:
      - Order
      summary: Delete Specific Order
      operationId: DeleteSpecificOrder
      parameters:
      - name: id
        in: path
        description: ''
        required: true
        style: simple
        schema:
          type: string
      responses:
        '200':
          description: ''
          headers: {}
      deprecated: false
      security:
      - bearer: []
...
components:
  schemas:
...
    AddCriminalDictionaryRequest:
      title: AddCriminalDictionaryRequest
      required:
      - title
      - decs
      type: object
      properties:
        title:
          type: string
        decs:
          type: string
      example:
        title: Test
        decs: ' Test Test Test Test Test Test Test Test Test '
    UpdateNewRequest:
      title: UpdateNewRequest
      required:
      - title
      type: object
      properties:
        title:
          type: string
      example:
        title: 'Test '
...
    AddOrderRequest:
      title: AddOrderRequest
      required:
      - reason
      - desc
      - phone
      - date
      type: object
      properties:
        reason:
          type: string
        desc:
          type: string
        phone:
          type: string
        date:
          type: string
      example:
        reason: Test
        desc: Test
        phone: '01544604708'
        date: 2012-7
...
  securitySchemes:
    bearer:
      type: http
      scheme: bearer
security: []
...